+++ examples/fp-ts-to-the-max-II.ts
@@ -1,14 +1,17 @@
-import { log } from '../src/Console'
-import { Kind, URIS } from '../src/HKT'
-import * as O from '../src/Option'
-import { randomInt } from '../src/Random'
-import * as T from '../src/Task'
+import * as assert from 'assert'
 import { createInterface } from 'readline'
-import * as S from '../src/State'
-import { Monad1 } from '../src/Monad'
-import { flow, pipe } from '../src/function'
+
 import { apS as apS_ } from '../src/Apply'
 import { chainFirst as chainFirst_ } from '../src/Chain'
+import { log } from '../src/Console'
+import { flow, pipe } from '../src/function'
+import { Kind, URIS } from '../src/HKT'
+import { Monad1 } from '../src/Monad'
+import * as O from '../src/Option'
+import { randomInt } from '../src/Random'
+import { append, dropLeft } from '../src/ReadonlyArray'
+import * as S from '../src/State'
+import * as T from '../src/Task'
 
 //
 // type classes
@@ -74,25 +77,25 @@ function parse(s: string): O.Option<number> {
 }
 
 function main<F extends URIS>(F: Main<F>): Kind<F, void> {
-  const chain =
+  const flatMap =
     <A, B>(f: (a: A) => Kind<F, B>) =>
     (ma: Kind<F, A>): Kind<F, B> =>
       F.chain(ma, f)
   const Do: Kind<F, {}> = F.of({})
   const apS = apS_(F)
-  const chainFirst = chainFirst_(F)
+  const tap = chainFirst_(F)
 
   // ask something and get the answer
   const ask = (question: string): Kind<F, string> =>
     pipe(
       F.putStrLn(question),
-      chain(() => F.getStrLn)
+      flatMap(() => F.getStrLn)
     )
 
   const shouldContinue = (name: string): Kind<F, boolean> => {
     return pipe(
       ask(`Do you want to continue, ${name} (y/n)?`),
-      chain((answer) => {
+      flatMap((answer) => {
         switch (answer.toLowerCase()) {
           case 'y':
             return F.of<boolean>(true)
@@ -110,7 +113,7 @@ function main<F extends URIS>(F: Main<F>): Kind<F, void> {
       Do,
       apS('secret', F.nextInt(5)),
       apS('guess', ask(`Dear ${name}, please guess a number from 1 to 5`)),
-      chain(({ secret, guess }) =>
+      flatMap(({ secret, guess }) =>
         pipe(
           parse(guess),
           O.fold(
@@ -122,15 +125,15 @@ function main<F extends URIS>(F: Main<F>): Kind<F, void> {
           )
         )
       ),
-      chain(() => shouldContinue(name)),
-      chain((b) => (b ? gameLoop(name) : F.of<void>(undefined)))
+      flatMap(() => shouldContinue(name)),
+      flatMap((b) => (b ? gameLoop(name) : F.of<void>(undefined)))
     )
   }
 
   return pipe(
     ask('What is your name?'),
-    chainFirst((name) => F.putStrLn(`Hello, ${name} welcome to the game!`)),
-    chain(gameLoop)
+    tap((name) => F.putStrLn(`Hello, ${name} welcome to the game!`)),
+    flatMap(gameLoop)
   )
 }
 
@@ -146,8 +149,6 @@ export const mainTask = main({
 //
 // tests
 //
-
-import { dropLeft, append } from '../src/ReadonlyArray'
 
 class TestData {
   constructor(
@@ -207,8 +208,6 @@ const mainTestTask = main({
 })
 
 const testExample = new TestData(['Giulio', '1', 'n'], [], [1])
-
-import * as assert from 'assert'
 
 assert.deepStrictEqual(mainTestTask(testExample), [
   undefined,