+++ scripts/linter.ts
@@ -402,12 +402,12 @@ export const parseFile = (src: ast.SourceFile): File => {
     src.getFunctions(),
     RA.filter((fd) => fd.isExported()),
     RA.map(parseFunctionDeclaration),
-    RA.concat(pipe(src.getInterfaces(), RA.chain(parseInterface))),
+    RA.concat(pipe(src.getInterfaces(), RA.flatMap(parseInterface))),
     RA.concat(
       pipe(
         src.getVariableDeclarations(),
         RA.filter((vd) => vd.isExported()),
-        RA.chain(parseVariableDeclaration)
+        RA.flatMap(parseVariableDeclaration)
       )
     )
   )
@@ -466,13 +466,13 @@ export const getTypeArguments = (type: Type): ReadonlyArray<string> => {
         type.typeArguments,
         RA.match(
           () => [type.name],
-          (types) => pipe([type.name], RA.concat(pipe(types, RA.chain(getTypeArguments))))
+          (types) => pipe([type.name], RA.concat(pipe(types, RA.flatMap(getTypeArguments))))
         )
       )
     case 'Signature': {
       const typeParameters = pipe(
         type.typeParameters,
-        RA.chain((tp) =>
+        RA.flatMap((tp) =>
           pipe(
             tp.constraint,
             O.map((t) => pipe(getTypeArguments(t), RA.prepend(tp.name))),
@@ -482,7 +482,7 @@ export const getTypeArguments = (type: Type): ReadonlyArray<string> => {
       )
       const typeArguments = pipe(
         type.parameters,
-        RA.chain((p) => getTypeArguments(p.type))
+        RA.flatMap((p) => getTypeArguments(p.type))
       )
       return pipe(typeParameters, RA.concat(typeArguments), RA.concat(getTypeArguments(type.returnType)))
     }
@@ -505,15 +505,15 @@ export const getTypeArguments = (type: Type): ReadonlyArray<string> => {
     case 'IndexedAccessType':
       return pipe(getTypeArguments(type.objectType), RA.concat(getTypeArguments(type.indexType)))
     case 'TupleType':
-      return pipe(type.elements, RA.chain(getTypeArguments))
+      return pipe(type.elements, RA.flatMap(getTypeArguments))
     case 'LiteralType':
     case 'Token':
       return RA.empty
     case 'UnionType':
     case 'IntersectionType':
-      return pipe(type.members, RA.chain(getTypeArguments))
+      return pipe(type.members, RA.flatMap(getTypeArguments))
     case 'Overloadings':
-      return pipe(type.signatures, RA.chain(getTypeArguments))
+      return pipe(type.signatures, RA.flatMap(getTypeArguments))
     case 'InferType':
       return getTypeArguments(type.typeParameter)
     case 'TypeParameterDeclaration':
@@ -535,14 +535,14 @@ export const lintType = (type: Type, path: string = string.empty): ReadonlyArray
     case 'Overloadings':
       return pipe(
         type.signatures,
-        RA.chainWithIndex((i, t) => lintType(t, `/${String(i)}`))
+        RA.flatMap((t, i) => lintType(t, `/${String(i)}`))
       )
     case 'TypeOperator':
       return lintType(type.type, path)
     case 'TupleType':
       return pipe(
         type.elements,
-        RA.chain((t) => lintType(t, path))
+        RA.flatMap((t) => lintType(t, path))
       )
     case 'IndexedAccessType':
       return pipe(lintType(type.objectType, path), RA.concat(lintType(type.indexType, path)))
@@ -553,7 +553,7 @@ export const lintType = (type: Type, path: string = string.empty): ReadonlyArray
     case 'IntersectionType':
       return pipe(
         type.members,
-        RA.chain((t) => lintType(t, path))
+        RA.flatMap((t) => lintType(t, path))
       )
     case 'ConditionalType':
       return pipe(
@@ -582,7 +582,7 @@ export const lintSignature = (path: string, s: Signature): ReadonlyArray<Lint> =
       ),
       pipe(
         s.parameters,
-        RA.chain((p) => getTypeArguments(p.type))
+        RA.flatMap((p) => getTypeArguments(p.type))
       )
     ),
     append(lintType(s.returnType))
@@ -592,7 +592,7 @@ export const lintSignature = (path: string, s: Signature): ReadonlyArray<Lint> =
 export const lintFunction = (filename: string, fd: FunctionDeclaration): ReadonlyArray<Lint> => {
   return pipe(
     fd.overloadings,
-    RA.chainWithIndex((i, s) => lintSignature(`${filename}/${fd.name}/${i}`, s))
+    RA.flatMap((s, i) => lintSignature(`${filename}/${fd.name}/${i}`, s))
   )
 }
 
@@ -600,7 +600,7 @@ export const lintFile = (file: File): ReadonlyArray<Lint> => {
   const functions = file.functions
   return pipe(
     functions,
-    RA.chain((f) => lintFunction(file.name, f))
+    RA.flatMap((f) => lintFunction(file.name, f))
   )
 }
 
@@ -640,7 +640,7 @@ paths.forEach((path) => project.addSourceFileAtPath(path))
 const files = pipe(project.getSourceFiles(), RA.map(parseFile))
 const checks = pipe(
   files,
-  RA.chain((f) => check(lintFile(f)))
+  RA.flatMap((f) => check(lintFile(f)))
 )
 if (checks.length > 0) {
   // tslint:disable-next-line: no-console